Protect answers routes.

+++ app/controllers/answers.rb
@@ -22,27 +22,41 @@ get '/questions/:id/answers/:answer_id/edit' do
   require_login
   @question = Question.find_by(id: params[:id])
   @answer = Answer.find_by(id: params[:answer_id])
-  erb :'/answers/edit'
-end
-
-put '/questions/:id/answers/:answer_id' do
-  @question = Question.find_by(id: params[:id])
-  @answer = Answer.find_by(id: params[:answer_id])
-  if @answer && @answer.update(params[:answer])
-    status 200
-    redirect "/questions/#{@question.id}"
+  if @answer.responder_id != current_user.id
+      erb :'error_404'
   else
-    status 400
-    @errors = @answer.errors.full_messages
     erb :'/answers/edit'
   end
 end
 
+put '/questions/:id/answers/:answer_id' do
+  require_login
+  @question = Question.find_by(id: params[:id])
+  @answer = Answer.find_by(id: params[:answer_id])
+  if @answer.responder_id != current_user.id
+      erb :'error_404'
+  else
+    if @answer && @answer.update(params[:answer])
+      status 200
+      redirect "/questions/#{@question.id}"
+    else
+      status 400
+      @errors = @answer.errors.full_messages
+      erb :'/answers/edit'
+    end
+  end
+end
+
 delete '/questions/:id/answers/:answer_id' do
+  require_login
   if answer = Answer.find_by(id: params[:answer_id])
-    answer.destroy
-    status 200
-    redirect "/questions/#{params[:id]}"
+    if answer.responder_id == current_user.id
+      answer.destroy
+      status 200
+      redirect "/questions/#{params[:id]}"
+    else
+      erb :'error_404'
+    end
   else
     status 400
     @errors = ['delete unsuccessful']   #I forget if answer.errors.full_messages works with this one. Will try later, but hard coding a fail message should work for now.
